# fix: update profile per fetch