Update profiler to better reflect times
```
Starting pprof server on :6060

=== Initial State ===
Alloc = 2 MiB
TotalAlloc = 3 MiB
Sys = 12 MiB
NumGC = 1

Total symbols to process: 500

=== Run 1/5 ===

=== Before Run 1 ===
Alloc = 116 MiB
TotalAlloc = 321 MiB
Sys = 222 MiB
NumGC = 16
Goroutines before fetch: 15

=== Run 1 Results ===
Successful: 500
Failed: 0
Total Duration: 25.249981125s
Fastest: 340.772125ms (AEVO/USDT:USDT)
Slowest: 25.227936542s (UFT/ETH)
Average: 12.805931894s
Average time per symbol: 50.499962ms
Goroutines after fetch: 1515

=== After Run 1 ===
Alloc = 119 MiB
TotalAlloc = 422 MiB
Sys = 332 MiB
NumGC = 17

=== Run 2/5 ===

=== Before Run 2 ===
Alloc = 97 MiB
TotalAlloc = 427 MiB
Sys = 332 MiB
NumGC = 19
Goroutines before fetch: 1515

=== Run 2 Results ===
Successful: 500
Failed: 0
Total Duration: 25.272242375s
Fastest: 294.679292ms (WAXP/BUSD)
Slowest: 25.250046625s (CHR/USDT:USDT)
Average: 12.76846172s
Average time per symbol: 50.544484ms
Goroutines after fetch: 3015

=== After Run 2 ===
Alloc = 108 MiB
TotalAlloc = 527 MiB
Sys = 347 MiB
NumGC = 20

=== Run 3/5 ===

=== Before Run 3 ===
Alloc = 98 MiB
TotalAlloc = 532 MiB
Sys = 347 MiB
NumGC = 22
Goroutines before fetch: 3015

=== Run 3 Results ===
Successful: 500
Failed: 0
Total Duration: 25.245445959s
Fastest: 296.74725ms (MLN/BNB)
Slowest: 25.224818708s (SNGLS/ETH)
Average: 12.771151634s
Average time per symbol: 50.490891ms
Goroutines after fetch: 4515

=== After Run 3 ===
Alloc = 107 MiB
TotalAlloc = 632 MiB
Sys = 359 MiB
NumGC = 23

=== Run 4/5 ===

=== Before Run 4 ===
Alloc = 99 MiB
TotalAlloc = 637 MiB
Sys = 359 MiB
NumGC = 25
Goroutines before fetch: 4515

=== Run 4 Results ===
Successful: 500
Failed: 0
Total Duration: 25.323823083s
Fastest: 329.482542ms (AEVO/USDT:USDT)
Slowest: 25.301990458s (EGLD/USD:EGLD)
Average: 12.771743831s
Average time per symbol: 50.647646ms
Goroutines after fetch: 6015

=== After Run 4 ===
Alloc = 105 MiB
TotalAlloc = 737 MiB
Sys = 368 MiB
NumGC = 26

=== Run 5/5 ===

=== Before Run 5 ===
Alloc = 100 MiB
TotalAlloc = 743 MiB
Sys = 368 MiB
NumGC = 28
Goroutines before fetch: 6015

=== Run 5 Results ===
Successful: 500
Failed: 0
Total Duration: 25.247352083s
Fastest: 296.191834ms (ALPACA/BTC)
Slowest: 25.224638542s (ETH/AUD)
Average: 12.770287225s
Average time per symbol: 50.494704ms
Goroutines after fetch: 7515

=== After Run 5 ===
Alloc = 105 MiB
TotalAlloc = 843 MiB
Sys = 377 MiB
NumGC = 29

=== Aggregate Results (5 runs) ===
Average Duration per Run: 25.267768925s
Success Rate: 100.00%

Timing Statistics:
Fastest Call: 294.679292ms (WAXP/BUSD)
Slowest Call: 25.301990458s (EGLD/USD:EGLD)
Average Call: 12.777515261s
Total Calls: 2500

=== Final State ===
Alloc = 12 MiB
TotalAlloc = 853 MiB
Sys = 377 MiB
NumGC = 32
```